Create Survey List
Create Survey List from Figma file.
The survey list page simply allows the Employer to see an overview of all of the surveys they have created. This list should be paginated to avoid infinite scrolling. When clicking the action icon on a particular survey, the employer should be navigated to the Results page. You can make use of https://tablericons.com/ for icons that are needed.
You can find the high‐fidelity mockup in Figma by following this
link: https://www.figma.com/file/DkGGwg5SUPDTokRo9WKZ8x/Survey‐Test