Create Responses page
The Responses page
Clicking on the action icon on the above screen should allow you to see responses from each individual submission. This should allow the employer to view the entire survey response for a selected individual. Unfortunately, the UI designer ran out of time to produce a design for this screen. Below is the wireframe. Please convert the wireframe into a UI screen using the same look and feel as The Survey List Page.